Refactor 404 page handling: remove old NotFoundPage component, update imports, and ensure proper token validation in ResetPasswordForm
